move auth fixture
